Add acceptance tests in core commands
